% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-events.R
\name{add_bednet_events}
\alias{add_bednet_events}
\title{Add bednet rounds}
\usage{
add_bednet_events(
  xds_obj,
  start_day,
  type = "pbo",
  event_length = 20,
  coverage = 1,
  contact = 1,
  shock = 1,
  d_50 = 365,
  d_shape = 1/365
)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{start_day}{the Julian start dates of bednet events}

\item{type}{the type of net used}

\item{event_length}{the number of days it took to spray the houses}

\item{coverage}{maximum effective coverage}

\item{contact}{maximum effective contact}

\item{shock}{maximum shock size}

\item{d_50}{half-coverage parameter}

\item{d_shape}{coverage shape parameter}
}
\value{
a \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
\keyword{internal}
