% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet.R
\name{Bed_Net_1}
\alias{Bed_Net_1}
\title{Bednet, Stage 1}
\usage{
Bed_Net_1(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the value of exogenous variables related to
bed nets
}
\keyword{internal}
