% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-access.R
\name{Bed_Net_Access.none}
\alias{Bed_Net_Access.none}
\title{Set no own_bednets}
\usage{
\method{Bed_Net_Access}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for own_bednets
}
\keyword{internal}
