% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-access.R
\name{Bed_Net_Access}
\alias{Bed_Net_Access}
\title{Compute Bed Net Access}
\usage{
Bed_Net_Access(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the value of bed net access
}
\keyword{internal}
