% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-func.R
\name{Bed_Net_Coverage.func}
\alias{Bed_Net_Coverage.func}
\title{Bed Net Coverage}
\usage{
\method{Bed_Net_Coverage}{func}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
This computes
\emph{effective_coverage}
}
