% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage.R
\name{Bed_Net_Coverage.none}
\alias{Bed_Net_Coverage.none}
\title{Set no cover_bednets}
\usage{
\method{Bed_Net_Coverage}{none}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for cover_bednets
}
\keyword{internal}
