% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-effect_sizes-lemenach.R
\name{Bed_Net_Effect_Sizes.lemenach}
\alias{Bed_Net_Effect_Sizes.lemenach}
\title{Modify baseline values due to vector control}
\usage{
\method{Bed_Net_Effect_Sizes}{lemenach}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a named \link{list}
}
\description{
Implements \link{Bed_Net_Effect_Sizes} for the Le Menach ITN model of vector control
}
\seealso{
\code{\link[=compute_bednet_effect_sizes_lemenach]{compute_bednet_effect_sizes_lemenach()}}
}
\keyword{internal}
