% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-effect_sizes.R
\name{Bed_Net_Effect_Sizes.none}
\alias{Bed_Net_Effect_Sizes.none}
\title{The null Model for Bed Net Effect Sizes}
\usage{
\method{Bed_Net_Effect_Sizes}{none}(t, y, xds_obj, s = 1)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the \strong{\code{xds}} model object
}
\description{
Do nothing
}
\keyword{internal}
