% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_hosts.R, R/other_blood_hosts.R
\name{blood_hosts_dynamics}
\alias{blood_hosts_dynamics}
\title{blood_hosts Dynamics}
\usage{
blood_hosts_dynamics(t, y, xds_obj)

blood_hosts_dynamics(t, y, xds_obj)
}
\arguments{
\item{t}{current time}

\item{y}{the state variable vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object

an \strong{\code{xds}} object
}
\description{
Compute and update blood host availability on the \code{XY_interface}.

Compute and update blood host availability on the \code{XY_interface}.
}
\keyword{internal}
