% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-multiround.R
\name{change_bednet_coverage_multiround}
\alias{change_bednet_coverage_multiround}
\title{Setup Muti-Round Bed Net coverage}
\usage{
change_bednet_coverage_multiround(xds_obj, peak_access)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{peak_access}{peak access}
}
\value{
an \strong{\code{xds}} model object
}
\description{
With multi-round bed net coverage, there
is a different relationship between coverage
and coverage in each round
}
