% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-coverage.R
\name{change_irs_coverage}
\alias{change_irs_coverage}
\title{Get irs coverage}
\usage{
change_irs_coverage(coverage, xds_obj)
}
\arguments{
\item{coverage}{coverage parameters}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
maximum irs coverage, as a vector
}
\description{
Get the values from a model
that set maximum irs coverage
}
\keyword{internal}
