% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-shock-multiround.R
\name{change_irs_shock_multiround}
\alias{change_irs_shock_multiround}
\title{Change IRS Shocks}
\usage{
change_irs_shock_multiround(xds_obj, shock)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{shock}{the scaling parameter}
}
\description{
For \code{eir} models, set up a multi-round
sharkbite function to model effect sizes
of IRS on the EIR as a perturbation.
}
\keyword{internal}
