% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_season.R
\name{change_season.Lambda}
\alias{change_season.Lambda}
\title{Change season parameters}
\usage{
\method{change_season}{Lambda}(X, xds_obj, ix = 1)
}
\arguments{
\item{X}{a list with new parameters for bottom, phase, and pw}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Change parameters for the seasonality function
when \code{forced_by = "Lambda"}
}
\keyword{internal}
