% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_season.R
\name{change_season.none}
\alias{change_season.none}
\title{Change season parameters}
\usage{
\method{change_season}{none}(X, xds_obj, s = 1)
}
\arguments{
\item{X}{a list with new parameters for bottom, phase, and pw}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
A short description...

Change parameters for the seasonality function
when \code{forced_by = "none"}
}
\keyword{internal}
