% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_season.R
\name{change_season}
\alias{change_season}
\title{Change season parameters}
\usage{
change_season(X, xds_obj, ix = 1)
}
\arguments{
\item{X}{a list with new parameters for bottom, phase, and pw}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{ix}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Change parameters for the seasonality function
}
\keyword{internal}
