% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{Development.func}
\alias{Development.func}
\title{Set no development}
\usage{
\method{Development}{func}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
[list]
}
\description{
The null model for development
}
