% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{Development}
\alias{Development}
\title{Set the development}
\usage{
Development(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
Set the value of exogenous variables related to
development
}
