% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcing.R
\name{dynamic_forcing}
\alias{dynamic_forcing}
\title{Set up dynamic forcing}
\usage{
dynamic_forcing(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
