% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation.R
\name{dynamic_importation.dynamic}
\alias{dynamic_importation.dynamic}
\title{Activate Dynamic Importation (dynamic)}
\usage{
\method{dynamic_importation}{dynamic}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
No-op: the \code{importation_obj} is already dynamic.
}
\keyword{internal}
