% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation.R
\name{dynamic_importation.none}
\alias{dynamic_importation.none}
\title{Activate Dynamic Importation (none)}
\usage{
\method{dynamic_importation}{none}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Upgrades the \code{importation_obj} from class \code{none} to
class \code{dynamic} and initialises the travel and visitor sub-objects.
}
\keyword{internal}
