% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{dynamic_resources.dynamic}
\alias{dynamic_resources.dynamic}
\title{Set up dynamic resources (already dynamic)}
\usage{
\method{dynamic_resources}{dynamic}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
No-op when resources are already dynamic.
}
\keyword{internal}
