% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{dynamic_resources.none}
\alias{dynamic_resources.none}
\title{Set up dynamic resources (none -> dynamic)}
\usage{
\method{dynamic_resources}{none}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Transitions from the default \code{none} class to \code{dynamic},
and initializes the sub-junction objects with \code{none} (no-op) defaults.
}
\keyword{internal}
