% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{dynamic_weather}
\alias{dynamic_weather}
\title{Set up dynamic weather}
\usage{
dynamic_weather(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic weather has not
already been set up, then turn on dynamic
weather and set up all its null models
}
\keyword{internal}
