% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{dynamic_weather.setup}
\alias{dynamic_weather.setup}
\title{Set up dynamic weather}
\usage{
\method{dynamic_weather}{setup}(pars)
}
\arguments{
\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
If dynamic weather has not
already been set up, then turn on dynamic
weather and set all the
}
