% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-contact.R
\name{get_bednet_contact}
\alias{get_bednet_contact}
\title{Get bednet coverage}
\usage{
get_bednet_contact(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Get the values from a model
that set maximum bednet coverage
}
\keyword{internal}
