% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_variables.splineX}
\alias{get_variables.splineX}
\title{Get Variables}
\usage{
\method{get_variables}{splineX}(opts, t, y, xds_obj, ix)
}
\arguments{
\item{opts}{an options list and dispatching}

\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{ix}{the species index}
}
\value{
a list of named variables
}
\description{
get_variables for \code{splineX}
}
\keyword{internal}
