% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{habitat_dynamics}
\alias{habitat_dynamics}
\title{Habitat Dynamics (internal)}
\usage{
habitat_dynamics(t, y, xds_obj)
}
\arguments{
\item{t}{current time}

\item{y}{state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Compute and update habitat search weights and bad habitat
availability on the \code{ML_interface}.
}
\keyword{internal}
