% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{HabitatDynamics.dynamic}
\alias{HabitatDynamics.dynamic}
\title{HabitatDynamics (dynamic)}
\usage{
\method{HabitatDynamics}{dynamic}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Dynamically updates habitat search weights and bad habitat availability.
}
\keyword{internal}
