% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{HabitatDynamics.setup}
\alias{HabitatDynamics.setup}
\title{HabitatDynamics (setup)}
\usage{
\method{HabitatDynamics}{setup}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Transitions habitat_obj from \code{setup} to \code{static} class on
the first call, triggering computation of egg-laying matrices.
}
\keyword{internal}
