% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{HabitatDynamics.static}
\alias{HabitatDynamics.static}
\title{HabitatDynamics (static)}
\usage{
\method{HabitatDynamics}{static}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
No-op for static habitat weights.
}
\keyword{internal}
