% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{Hydrology.func}
\alias{Hydrology.func}
\title{Set no hydrology}
\usage{
\method{Hydrology}{func}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for hydrology
}
