% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation.R
\name{Importation.dynamic}
\alias{Importation.dynamic}
\title{Importation (dynamic)}
\usage{
\method{Importation}{dynamic}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Dynamic method for the \code{Importation} junction. Calls
\link{Travel} and \link{Visitors} to update time at home, travel EIR,
visitor availability, and visitor infectiousness.
}
\keyword{internal}
