% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs.R
\name{IRS_1.dynamic}
\alias{IRS_1.dynamic}
\title{Set the irs}
\usage{
\method{IRS_1}{dynamic}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the value of exogenous variables related to
irs
}
\keyword{internal}
