% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-contact-multiround.R
\name{IRS_Contact.multiround}
\alias{IRS_Contact.multiround}
\title{IRS Contact for Multiround Models}
\usage{
\method{IRS_Contact}{multiround}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
A model for IRS contact over time
when there have been several rounds of IRS
}
\keyword{internal}
