% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-contact.R
\name{IRS_Contact}
\alias{IRS_Contact}
\title{IRS Contact}
\usage{
IRS_Contact(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Compute IRS contact
}
\keyword{internal}
