% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-coverage-func.R
\name{IRS_Coverage.func}
\alias{IRS_Coverage.func}
\title{Set no irs_coverage}
\usage{
\method{IRS_Coverage}{func}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for irs_coverage
}
\keyword{internal}
