% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-coverage-multiround.R
\name{IRS_Coverage.multiround}
\alias{IRS_Coverage.multiround}
\title{IRS Coverage with Multiple Rounds}
\usage{
\method{IRS_Coverage}{multiround}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
A model for IRS coverage over time
when there have been several, different IRS models
}
\keyword{internal}
