% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-coverage.R
\name{IRS_Coverage}
\alias{IRS_Coverage}
\title{IRS Coverage}
\usage{
IRS_Coverage(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Compute IRS coverage
}
\keyword{internal}
