% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effect_sizes.R
\name{IRS_Effect_Sizes.none}
\alias{IRS_Effect_Sizes.none}
\title{Set no use_bednets}
\usage{
\method{IRS_Effect_Sizes}{none}(t, y, xds_obj, s = 1)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
\link{list}
}
\description{
The null model for use_bednets
}
\keyword{internal}
