% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effect_sizes-simple.R
\name{IRS_Effect_Sizes.simple}
\alias{IRS_Effect_Sizes.simple}
\title{Modify baseline values due to vector control}
\usage{
\method{IRS_Effect_Sizes}{simple}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
a named \link{list}
}
\description{
Implements \link{IRS_Effect_Sizes} for the Le Menach IRS model of vector control
}
\seealso{
\code{\link[=compute_irs_effect_sizes_simple]{compute_irs_effect_sizes_simple()}}
}
\keyword{internal}
