% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effects.R
\name{IRS_Effects}
\alias{IRS_Effects}
\title{IRS Effects}
\usage{
IRS_Effects(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the value of variables to model
the effects of IRS
}
\keyword{internal}
