% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-effect_sizes-lemenach.R
\name{make_bednet_effect_sizes_lemenach}
\alias{make_bednet_effect_sizes_lemenach}
\title{Set up dynamic forcing}
\usage{
make_bednet_effect_sizes_lemenach(
  options = list(),
  tau0_frac = c(0.68/3, 2.32/3),
  rr = 0.56,
  ss = 0.03
)
}
\arguments{
\item{options}{a list of options to override defaults}

\item{tau0_frac}{a \link{numeric} vector giving the proportion of a feeding cycle spent
host seeking/bloodfeeding \emph{vs.} resting/oviposition}

\item{rr}{probability of mosquito being repelled upon contact with ITN}

\item{ss}{probability of mosquito successfully feeding upon contact with ITN}
}
\value{
a bed net effect size model object
}
\description{
Set up the model for the effect sizes of insecticide treated nets
using the model by Le Menach, \emph{et al.}
}
\keyword{internal}
