% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-profiles.R
\name{make_bednet_round}
\alias{make_bednet_round}
\title{Make an Bed Net Round}
\usage{
make_bednet_round(d_50, d_shape, start_day, peak, length = 20, pw = 1)
}
\arguments{
\item{d_50}{the half-saturation day}

\item{d_shape}{the shape}

\item{start_day}{the start day for the bednet round}

\item{peak}{the maximum value}

\item{length}{the number of days it took to complete spraying}

\item{pw}{a shape parameter}
}
\value{
a \code{sharkfin} function object
}
\description{
Return the parameters
to make sharkfin function for a bed net round with one of the
effective coverage profiles in the \code{bednet_profiles} table.

The parameters
specify a model for the
"effective coverage" after a mass bed net distribution
}
\keyword{internal}
