% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-profiles.R
\name{make_bednet_shock}
\alias{make_bednet_shock}
\title{Make a Bed Net Round}
\usage{
make_bednet_shock(d_50, d_shape, start_day, peak, length = 20, pw = 1)
}
\arguments{
\item{d_50}{the half-saturation day}

\item{d_shape}{the shape}

\item{start_day}{the start day for the bednet round}

\item{peak}{the maximum value}

\item{length}{the number of days it took to complete spraying}

\item{pw}{a shape parameter}
}
\value{
a \code{sharkbite} function object
}
\description{
Return the parameters
to make sharkfin function for an bednet round with one of the
pre-defined profiles in the \code{bednet_profiles} table.

The parameters
specify a model for bed net coverage
after a mass distribution
}
\keyword{internal}
