% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_treatment.R
\name{make_F_mass_treat}
\alias{make_F_mass_treat}
\title{Set up dynamic forcing}
\usage{
make_F_mass_treat(treat)
}
\arguments{
\item{treat}{a list with parameters for all the rounds}
}
\value{
a \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
\keyword{internal}
