% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-contact-func.R
\name{make_irs_contact_func}
\alias{make_irs_contact_func}
\title{Set up dynamic forcing}
\usage{
make_irs_contact_func(
  options = list(),
  mean = 1,
  F_season = F_one,
  season_par = list(),
  F_trend = F_one,
  trend_par = list()
)
}
\arguments{
\item{options}{a list of options to override defaults}

\item{mean}{the mean irs_contact}

\item{F_season}{the seasonal signal in irs contact}

\item{season_par}{parameters to configure F_season}

\item{F_trend}{a temporal trend in irs contact}

\item{trend_par}{parameters to configure F_trend}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
