% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-rounds.R
\name{make_irs_efsz_profile}
\alias{make_irs_efsz_profile}
\title{Make an IRS Effect Size Curve}
\usage{
make_irs_efsz_profile(d_50, d_shape, start_day, peak, elength = 20, pw = 1)
}
\arguments{
\item{d_50}{the day when efficacy reaches 50\%}

\item{d_shape}{the decay shape}

\item{start_day}{the start day for the IRS round}

\item{peak}{the maximum value}

\item{elength}{the number of days it took to complete spraying}

\item{pw}{a shape parameter}
}
\value{
a \code{sharkbite} function object
}
\description{
Return the parameters
to make sharkfin function for an irs round.
}
\keyword{internal}
