% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-rounds.R
\name{make_irs_shock}
\alias{make_irs_shock}
\title{Make an IRS Round}
\usage{
make_irs_shock(irs_type, start_day, peak, elength = 20, pw = 1)
}
\arguments{
\item{irs_type}{the name of the IRS type}

\item{start_day}{the start day for the IRS round}

\item{peak}{the maximum value}

\item{elength}{the number of days it took to complete spraying}

\item{pw}{a shape parameter}
}
\value{
a \code{sharkbite} function object
}
\description{
Return the parameters
to make sharkfin function for an irs round with one of the
pesticides in the \code{irs_profiles} table.

The parameters
specify a model for the
"killing potential" from the start of the spray round through the end.
}
\keyword{internal}
