% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_none_object}
\alias{make_none_object}
\title{Make "none" object}
\usage{
make_none_object(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Make a "none" object
}
\keyword{internal}
