% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiround.R
\name{makepar_F_multiround}
\alias{makepar_F_multiround}
\title{Make Parameters for a Multiround Function}
\usage{
makepar_F_multiround(nRounds, rounds)
}
\arguments{
\item{nRounds}{the rate parameter}

\item{rounds}{the rounds}
}
\value{
a multiround function
}
\description{
Return an object to configure
multiple \code{sharkfin} functions representing
coverage levels over many mass distribution events
}
\keyword{internal}
