% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multishock.R
\name{makepar_F_multishock}
\alias{makepar_F_multishock}
\title{Make Parameters for a multishock Function}
\usage{
makepar_F_multishock(nRounds, rounds)
}
\arguments{
\item{nRounds}{the rate parameter}

\item{rounds}{the rounds}
}
\value{
a multishock function
}
\description{
Return an object to configure
multiple \code{sharkbite} functions representing
coverage levels over many mass distribution events
}
\keyword{internal}
