% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_hosts.R, R/other_blood_hosts.R
\name{OtherBloodHosts.dynamic}
\alias{OtherBloodHosts.dynamic}
\title{Blood Hosts (dynamic)}
\usage{
\method{OtherBloodHosts}{dynamic}(t, y, xds_obj)

\method{OtherBloodHosts}{dynamic}(t, y, xds_obj)
}
\arguments{
\item{t}{time}

\item{y}{state variables vector}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object

an \strong{\code{xds}} object
}
\description{
Dynamically updates blood host availability.

Dynamically updates blood host availability.
}
\keyword{internal}
