% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_hosts.R
\name{OtherBloodHosts}
\alias{OtherBloodHosts}
\title{Blood Hosts (static)}
\usage{
OtherBloodHosts(t, y, xds_obj)
}
\arguments{
\item{t}{time}

\item{y}{state variables vector}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
For a static model for blood hosts, the function
does not update anything.
}
\keyword{internal}
