% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_hosts.R, R/other_blood_hosts.R
\name{OtherBloodHosts.setup}
\alias{OtherBloodHosts.setup}
\title{Blood Hosts (setup)}
\usage{
\method{OtherBloodHosts}{setup}(t, y, xds_obj)

\method{OtherBloodHosts}{setup}(t, y, xds_obj)
}
\arguments{
\item{t}{time}

\item{y}{state variables vector}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object

an \strong{\code{xds}} object
}
\description{
Sets a static value for blood host availability, resets the class
to \code{static}, and triggers an update to the blood feeding model.

Sets a static value for blood host availability, resets the class
to \code{static}, and triggers an update to the blood feeding model.
}
\keyword{internal}
