% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovitrap.R
\name{OviTraps.none}
\alias{OviTraps.none}
\title{Methods for oviposition traps}
\usage{
\method{OviTraps}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a [list]}
}
\value{
[list]
}
\description{
Implements [OviTraps] for the none model (do nothing)
}
