% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{Rainfall.none}
\alias{Rainfall.none}
\title{Set no rainfall}
\usage{
\method{Rainfall}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for rainfall
}
