% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{Rainfall}
\alias{Rainfall}
\title{Set the rainfall}
\usage{
Rainfall(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
Set the value of exogenous variables related to
rainfall
}
