% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readXDS}
\alias{readXDS}
\title{readRDS for \code{xds} Objects}
\usage{
readXDS(filename)
}
\arguments{
\item{filename}{the file name}
}
\value{
an \strong{\code{xds}} object
}
\description{
Read the \code{xds} object using \code{readRDS} and rebuild the forcing functions
}
\keyword{internal}
