% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{Resources.dynamic}
\alias{Resources.dynamic}
\title{Dynamic Resources}
\usage{
\method{Resources}{dynamic}(t, y, xds_obj)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Compute availability of resources affecting adult mosquito
bionomics: blood hosts, aquatic habitats, traps, and sugar.
This is the dynamic implementation of the \link[ramp.xds:Resources]{ramp.xds::Resources} junction.
\itemize{
\item \link{OtherBloodHosts} availability of alternative blood hosts
\item \link{HabitatDynamics} to modify habitat search weights
\item \link{Traps} to compute availability of oviposition traps
\item \link{Sugar} availability of sugar
}
}
\seealso{
\link{dynamic_resources}
}
\keyword{internal}
