% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{Resources.forced}
\alias{Resources.forced}
\title{Methods for resources}
\usage{
Resources.forced(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
[list]
}
\description{
Implements [Resources]
}
