% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-access.R
\name{setup_bednet_access.none}
\alias{setup_bednet_access.none}
\title{Set Up the Null Bed Net Access Object}
\usage{
\method{setup_bednet_access}{none}(name, xds_obj, options = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{options}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up the null bed net ownership object
}
\keyword{internal}
