% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-contact-multiround.R
\name{setup_bednet_contact.multiround}
\alias{setup_bednet_contact.multiround}
\title{Setup Muti-Round Bed Net Contact}
\usage{
\method{setup_bednet_contact}{multiround}(name = "multiround", xds_obj, options = list())
}
\arguments{
\item{name}{the model name}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{options}{a list of options to override defaults}
}
\description{
With multi-round bed net contact, there
is a different relationship between contact
and contact in each round
}
\keyword{internal}
