% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-func.R
\name{setup_bednet_coverage.func}
\alias{setup_bednet_coverage.func}
\title{Set Up a Bed Net Coverage Function}
\usage{
\method{setup_bednet_coverage}{func}(name = "func", xds_obj, options = list())
}
\arguments{
\item{name}{the model name}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{options}{a list of options to override defaults}
}
\description{
This sets up a function to model bed net coverage.
\itemize{
\item A function F_cover
}
}
\keyword{internal}
