% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-multiround.R
\name{setup_bednet_coverage.multiround}
\alias{setup_bednet_coverage.multiround}
\title{Set Up a Bed Net Coverage Function}
\usage{
\method{setup_bednet_coverage}{multiround}(name = "multiround", xds_obj, options = list())
}
\arguments{
\item{name}{the model name}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{options}{a list of options to override defaults}
}
\value{
an \strong{\code{xds}} model object
}
\description{
This sets up a function to set the value of bed net
coverage
}
