% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage.R
\name{setup_bednet_coverage}
\alias{setup_bednet_coverage}
\title{Set Up the Bed Net Coverage Object}
\usage{
setup_bednet_coverage(name, xds_obj, options = list())
}
\arguments{
\item{name}{the model name}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{options}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up an object to model
bed net effective coverage, which is defined
as the product of access and use.
}
\keyword{internal}
