% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-effect_sizes-lemenach.R
\name{setup_bednet_effect_sizes.lemenach}
\alias{setup_bednet_effect_sizes.lemenach}
\title{Set up dynamic forcing}
\usage{
\method{setup_bednet_effect_sizes}{lemenach}(name, xds_obj, s = 1, options = list())
}
\arguments{
\item{name}{the bed net effect sizes model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a list of options to override defaults}
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
\keyword{internal}
