% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-effect_sizes.R
\name{setup_bednet_effect_sizes.none}
\alias{setup_bednet_effect_sizes.none}
\title{Set Up the Null Model for Bed Net Effect Sizes}
\usage{
\method{setup_bednet_effect_sizes}{none}(name, xds_obj, s = 1, options = list())
}
\arguments{
\item{name}{the bed net effect sizes model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a list of options to override defaults}
}
\value{
the \strong{\code{xds}} model object
}
\description{
Set up the null bed net effect sizes model object
}
\keyword{internal}
