% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-use.R
\name{setup_bednet_use.none}
\alias{setup_bednet_use.none}
\title{Set Up the Null Bed Net Use Object}
\usage{
\method{setup_bednet_use}{none}(name, xds_obj, options = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{a list of options to override defaults}
}
\value{
the \strong{\code{xds}} model object
}
\description{
Set up the null bed net use object
}
\keyword{internal}
