% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-use.R
\name{setup_bednet_use}
\alias{setup_bednet_use}
\title{Set Up the Bed Net Use Object}
\usage{
setup_bednet_use(name, xds_obj, options = list())
}
\arguments{
\item{name}{the model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{a list of options to override defaults}
}
\value{
the \strong{\code{xds}} model object
}
\description{
Set up an object to model
bed net use
}
\keyword{internal}
