% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet.R
\name{setup_bednets}
\alias{setup_bednets}
\title{Set Up a Bed Net Model}
\usage{
setup_bednets(
  xds_obj,
  access_name = "none",
  access_opts = list(),
  use_name = "none",
  use_opts = list(),
  effects_name = "none",
  effects_opts = list(),
  coverage_name = "none",
  coverage_opts = list(),
  contact_name = "linear",
  contact_opts = list(cp = 1),
  effect_sizes_name = "none",
  effect_sizes_opts = list()
)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{access_name}{the name of a model for bed net access}

\item{access_opts}{options for the bed net access model}

\item{use_name}{the name of a model for bed net usage}

\item{use_opts}{options for the bed net usage model}

\item{effects_name}{the name of a model for bed net effects}

\item{effects_opts}{options for the bed net effects model}

\item{coverage_name}{the name of a model for bed net coverage}

\item{coverage_opts}{options for the bed net coverage model}

\item{contact_name}{the name of a model for bed net contact}

\item{contact_opts}{options for the bed net contact model}

\item{effect_sizes_name}{the name of a model for bed net effect sizes}

\item{effect_sizes_opts}{options for the bed net effect sizes model}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
This sets up a bed net
model
}
\keyword{internal}
