% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_hosts.R, R/other_blood_hosts.R
\name{setup_blood_host_object}
\alias{setup_blood_host_object}
\title{Setup the Blood Hosts Object}
\usage{
setup_blood_host_object(xds_obj)

setup_blood_host_object(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object

an \strong{\code{xds}} object
}
\description{
Setup the object to handle
other blood hosts -- potential blood hosts
that are not hosts for the pathogen

Setup the object to handle
other blood hosts -- potential blood hosts
that are not hosts for the pathogen
}
\keyword{internal}
