% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_hosts.R, R/other_blood_hosts.R
\name{setup_blood_hosts}
\alias{setup_blood_hosts}
\title{Set up blood hosts model}
\usage{
setup_blood_hosts(mod_name, xds_obj, s, options)

setup_blood_hosts(mod_name, xds_obj, s, options)
}
\arguments{
\item{mod_name}{the model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a list to configure \code{F_blood_host}}
}
\value{
an \strong{\code{xds}} object

an \strong{\code{xds}} object
}
\description{
Setup a model for blood host availability.

Setup a model for blood host availability.
}
\keyword{internal}
