% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_hosts.R, R/other_blood_hosts.R
\name{setup_blood_hosts.ts_func}
\alias{setup_blood_hosts.ts_func}
\title{Set up blood hosts model (ts_func)}
\usage{
\method{setup_blood_hosts}{ts_func}(mod_name, xds_obj, s, options)

\method{setup_blood_hosts}{ts_func}(mod_name, xds_obj, s, options)
}
\arguments{
\item{mod_name}{the model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a list to configure \code{F_blood_host}}
}
\value{
an \strong{\code{xds}} object

an \strong{\code{xds}} object
}
\description{
Setup a time series function for blood host availability.

Setup a time series function for blood host availability.
}
\keyword{internal}
