% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-contact-multiround.R
\name{setup_F_contact_bednet_multiround}
\alias{setup_F_contact_bednet_multiround}
\title{Make \code{F_contact} for bednet}
\usage{
setup_F_contact_bednet_multiround(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
set up the rounds
}
\description{
Set up the bednet rounds
}
\keyword{internal}
