% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-multiround.R
\name{setup_F_coverage_bednet_multiround}
\alias{setup_F_coverage_bednet_multiround}
\title{Make \code{F_coverage} for bednet}
\usage{
setup_F_coverage_bednet_multiround(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Set up the bednet rounds
}
