% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{setup_F_habitat_weights.ts_func}
\alias{setup_F_habitat_weights.ts_func}
\title{Setup Bad Habitat Availability}
\usage{
\method{setup_F_habitat_weights}{ts_func}(mod_name, xds_obj, s, options = list())
}
\arguments{
\item{mod_name}{the model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a list of setup options (see \link[ramp.func:make_ts_function]{ramp.func::make_ts_function})}
}
\value{
an \strong{\code{xds}} object
}
\description{
Setup a the function model for no time spent searching
and no exposure while searching
}
\keyword{internal}
