% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{setup_F_travel.ts_func}
\alias{setup_F_travel.ts_func}
\title{Set up travel as a time series function}
\usage{
\method{setup_F_travel}{ts_func}(setup_name, xds_obj, i, options = list())
}
\arguments{
\item{setup_name}{a string to dispatch \code{setup_F_travel}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{setup options for \code{F_travel}}
}
\value{
an \strong{\code{xds}} object
}
\description{
Setup a time series function for time spent traveling.
}
\keyword{internal}
