% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{setup_F_travel_eir}
\alias{setup_F_travel_eir}
\title{Setup the Travel EIR model}
\usage{
setup_F_travel_eir(setup_name, xds_obj, i, options = list())
}
\arguments{
\item{setup_name}{a string to dispatch \code{setup_F_travel_eir}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{a named list to set up \code{F_travel_eir}}
}
\value{
an \strong{\code{xds}} object
}
\description{
Setup a model for the dEIR experienced while traveling.
}
\keyword{internal}
