% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{setup_F_travel_eir.ts_func}
\alias{setup_F_travel_eir.ts_func}
\title{Set up travel EIR as a time series function}
\usage{
\method{setup_F_travel_eir}{ts_func}(setup_name, xds_obj, i, options = list())
}
\arguments{
\item{setup_name}{a string to dispatch \code{setup_F_travel_eir}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{a named list to set up \code{F_travel_eir}}
}
\value{
an \strong{\code{xds}} object
}
\description{
Setup a time series function for exposure while traveling.
}
\keyword{internal}
