% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{setup_F_vis_kappa}
\alias{setup_F_vis_kappa}
\title{Set up the visitor infectiousness model}
\usage{
setup_F_vis_kappa(mod_name, xds_obj, s, options)
}
\arguments{
\item{mod_name}{the model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a list to configure \code{F_vis_kappa}}
}
\value{
an \strong{\code{xds}} object
}
\description{
Setup a model for the net infectiousness of visitors.
}
\keyword{internal}
