% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{setup_F_visitors}
\alias{setup_F_visitors}
\title{Set up the visitors model}
\usage{
setup_F_visitors(mod_name, xds_obj, s, options)
}
\arguments{
\item{mod_name}{the model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a named list to setup \code{F_visitors}}
}
\value{
an \strong{\code{xds}} object
}
\description{
Setup a model for visitor availability.
}
