% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{setup_F_visitors.ts_func}
\alias{setup_F_visitors.ts_func}
\title{Set up visitor availability as a time series function}
\usage{
\method{setup_F_visitors}{ts_func}(mod_name, xds_obj, s, options)
}
\arguments{
\item{mod_name}{the model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a named list to setup \code{F_visitors}}
}
\value{
an \strong{\code{xds}} object
}
\description{
Setup a time series function for visitor availability.
}
\keyword{internal}
