% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{setup_habitat_dynamics_static}
\alias{setup_habitat_dynamics_static}
\title{Setup the egg laying object}
\usage{
setup_habitat_dynamics_static(pars)
}
\arguments{
\item{pars}{the model object}
}
\value{
a [list] vector
}
\description{
Sets up the egg-deposition matrix calU for the s^th species
}
