% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{setup_humidity_func}
\alias{setup_humidity_func}
\title{Set up dynamic forcing}
\usage{
setup_humidity_func(
  xds_obj,
  Topts = list(),
  mean = 80,
  F_season = F_one,
  F_trend = F_one
)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{Topts}{a list of options to override defaults}

\item{mean}{the mean humidity}

\item{F_season}{the seasonal signal in humidity}

\item{F_trend}{a temporal trend in humidity}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
