% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs.R
\name{setup_irs}
\alias{setup_irs}
\title{Set Up a Bed Net Model}
\usage{
setup_irs(
  xds_obj,
  spray_houses_name = "none",
  spray_houses_opts = list(),
  effects_name = "none",
  effects_opts = list(),
  coverage_name = "none",
  coverage_opts = list(),
  contact_name = "linear",
  contact_opts = list(cp = 1),
  effect_sizes_name = "none",
  effect_sizes_opts = list()
)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{spray_houses_name}{the name of a model for bed net ownership}

\item{spray_houses_opts}{options for the bed net ownership model}

\item{effects_name}{the name of a model for bed net effects}

\item{effects_opts}{options for the bed net effects model}

\item{coverage_name}{the name of a model for bed net coverage}

\item{coverage_opts}{options for the bed net coverage model}

\item{contact_name}{the name of a model for bed net contact}

\item{contact_opts}{options for the bed net contact model}

\item{effect_sizes_name}{the name of a model for bed net effect sizes}

\item{effect_sizes_opts}{options for the bed net effect sizes model}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
This sets up a bed net
model
}
\keyword{internal}
