% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-coverage-func.R
\name{setup_irs_coverage.func}
\alias{setup_irs_coverage.func}
\title{Set Up a IRS Coverage Function}
\usage{
\method{setup_irs_coverage}{func}(name = "func", xds_obj, options = list())
}
\arguments{
\item{name}{the model name}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{options}{a list of options to override defaults}
}
\description{
This sets up a function to set the value of bed net
coverage
}
\keyword{internal}
