% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effect_sizes.R
\name{setup_irs_effect_sizes.none}
\alias{setup_irs_effect_sizes.none}
\title{Set Up the Null IRS effect_sizes Object}
\usage{
\method{setup_irs_effect_sizes}{none}(name, xds_obj, s = 1, options = list())
}
\arguments{
\item{name}{the model name}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{s}{the vector species index}

\item{options}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up a model for IRS effect_sizes that does nothing.
}
\keyword{internal}
