% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effect_sizes.R
\name{setup_irs_effect_sizes}
\alias{setup_irs_effect_sizes}
\title{Set Up the IRS effect_sizes Object}
\usage{
setup_irs_effect_sizes(name, xds_obj, s = 1, options = list())
}
\arguments{
\item{name}{the model name}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{s}{the vector species index}

\item{options}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up an object to model
the effect_sizes of IRS, usually through the
computation of some intermediate variable
}
\keyword{internal}
