% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effects.R
\name{setup_irs_effects.none}
\alias{setup_irs_effects.none}
\title{Set Up the Null IRS Effects Object}
\usage{
\method{setup_irs_effects}{none}(name, xds_obj, options = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{options}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up a model for IRS effects that does nothing.
}
\keyword{internal}
