% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs.R
\name{setup_irs_object}
\alias{setup_irs_object}
\title{Set Up the Null IRS Model Object}
\usage{
setup_irs_object(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Sets up the null IRS model object
}
\keyword{internal}
