% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_treatment.R
\name{setup_mass_treat_multiround}
\alias{setup_mass_treat_multiround}
\title{Set Up a Bed Net Function}
\usage{
setup_mass_treat_multiround(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}
}
\description{
This sets up a function to set the value of bed net
treatage
}
\keyword{internal}
