% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{setup_no_humidity}
\alias{setup_no_humidity}
\title{Set up "no humidity"}
\usage{
setup_no_humidity(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up "no humidity"
}
