% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shock.R
\name{setup_no_shock}
\alias{setup_no_shock}
\title{Set up "no shock"}
\usage{
setup_no_shock(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up "no shock"
}
