% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{setup_no_weather}
\alias{setup_no_weather}
\title{Set up the no_forcing model for weather}
\usage{
setup_no_weather(xds_obj)
}
\arguments{
\item{xds_obj}{a [list]}
}
\value{
[list]
}
\description{
Set up the no_forcing model for weather
}
