% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{setup_resources_forced}
\alias{setup_resources_forced}
\title{Set up parameters for the static model for resource availability}
\usage{
setup_resources_forced(pars)
}
\arguments{
\item{pars}{a [list]}
}
\value{
none
}
\description{
Set up parameters for the static model for resource availability
}
