% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_season.R
\name{setup_season}
\alias{setup_season}
\title{Setup Seasonal Forcing}
\usage{
setup_season(X, xds_obj, ix = 1)
}
\arguments{
\item{X}{a seasonality function object}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{ix}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Setup seasonal forcing
}
\keyword{internal}
