% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shock.R
\name{setup_shock}
\alias{setup_shock}
\title{Set up dynamic forcing}
\usage{
setup_shock(name, pars, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an **`xds`** object}

\item{opts}{a list of options to override defaults}
}
\value{
an **`xds`** object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
