% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shock.R
\name{setup_shock_func}
\alias{setup_shock_func}
\title{Set up dynamic forcing}
\usage{
setup_shock_func(pars, opts = list(), eventT = 365, F_shock = NULL)
}
\arguments{
\item{pars}{an **`xds`** object}

\item{opts}{a list of options to override defaults}

\item{eventT}{the time when a shock occurs}

\item{F_shock}{the effects of the shock}
}
\value{
an **`xds`** object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
