% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-spray_houses.R
\name{setup_spray_houses.none}
\alias{setup_spray_houses.none}
\title{Set Up the Null House Spraying Object}
\usage{
\method{setup_spray_houses}{none}(name, xds_obj, options = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{options}{a list of options to override defaults}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set up the null house spraying model object
}
\keyword{internal}
