% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{setup_sugar_object}
\alias{setup_sugar_object}
\title{Setup the sugar junction object}
\usage{
setup_sugar_object(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Initializes \code{resources_obj$sugar_obj} with class \code{none},
ready for dynamic sugar configuration.
Called by \link{dynamic_resources}.
}
\keyword{internal}
