% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traps.R
\name{setup_traps_object}
\alias{setup_traps_object}
\title{Setup the traps junction object}
\usage{
setup_traps_object(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Initializes \code{resources_obj$traps_obj} with class \code{none},
ready for dynamic trap configuration.
Called by \link{dynamic_resources}.
}
\keyword{internal}
