% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{setup_travel_object}
\alias{setup_travel_object}
\title{Setup the Travel Sub-Object}
\usage{
setup_travel_object(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Setup the port sub-object to handle
travel dynamics. Initialises \code{travel_obj} on the \code{XY_interface}
and sets static (zero) defaults for \code{F_travel} and \code{F_travel_eir}.
}
\note{
The parameters \code{time_at_home} and \code{travel_EIR} are
initialised by \link[ramp.xds:setup_importation_object]{ramp.xds::setup_importation_object}; this function
only sets up the port machinery.
}
\keyword{internal}
