% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{setup_vector_control.dynamic}
\alias{setup_vector_control.dynamic}
\title{Set Up Vector Control}
\usage{
\method{setup_vector_control}{dynamic}(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
If
\code{class(xds_obj$vector_control_obj) == 'none'}
this function sets up the
objects for all implemented modes of vector control.
}
\keyword{internal}
