% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{setup_visitor_object}
\alias{setup_visitor_object}
\title{Setup the Visitor Sub-Object}
\usage{
setup_visitor_object(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Setup the port sub-object to handle visitor dynamics.
Initialises \code{visitor_obj} on the \code{XY_interface} and sets static
(zero) defaults for \code{F_visitors} and \code{F_vis_kappa}.
}
\note{
The parameters \code{visitors} and \code{vis_kappa} are initialised by
\link[ramp.xds:setup_importation_object]{ramp.xds::setup_importation_object}; this function only sets up the
port machinery.
}
\keyword{internal}
