% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{setup_visitors_static}
\alias{setup_visitors_static}
\title{Make parameters for the static model visitors (no visitors)}
\usage{
setup_visitors_static(pars)
}
\arguments{
\item{pars}{a [list]}
}
\value{
[list]
}
\description{
Make parameters for the static model visitors (no visitors)
}
