% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-profiles.R
\name{show_bednet_response_timeline}
\alias{show_bednet_response_timeline}
\title{Make Parameters for an bednet Round}
\usage{
show_bednet_response_timeline(d_50, d_shape, clr = "black", add = FALSE)
}
\arguments{
\item{d_50}{the half-saturation parameter}

\item{d_shape}{the shape parameter}

\item{clr}{the line color}

\item{add}{if TRUE, add to an existing plot}
}
\value{
a \strong{\code{xds}} object
}
\description{
Return the parameters
to call a sharkfin function to model
a single round of bednet.
}
\keyword{internal}
