% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-rounds.R
\name{show_irs_profile}
\alias{show_irs_profile}
\title{Show an IRS Profile}
\usage{
show_irs_profile(irs_type)
}
\arguments{
\item{irs_type}{the name of the IRS type}
}
\value{
a \strong{\code{xds}} object
}
\description{
Return the parameters
to call a sharkfin function to model
a single round of IRS.
}
\keyword{internal}
