% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-shock-multiround.R
\name{show_irs_shock}
\alias{show_irs_shock}
\title{Plot IRS Coverage}
\usage{
show_irs_shock(tt, xds_obj, clr = "black", add = FALSE)
}
\arguments{
\item{tt}{a set of time points}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{clr}{plotting color}

\item{add}{add to an existing plot}
}
\value{
a \strong{\code{ramp.xds}} xds_obj object
}
\description{
Plot IRS Coverage
}
\keyword{internal}
