% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{show_spline_points}
\alias{show_spline_points}
\title{Show Events}
\usage{
show_spline_points(xds_obj, yval = 0)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}}  model object}

\item{yval}{the position of the interpolation points}
}
\value{
a \strong{\code{ramp.xds}}  model object
}
\description{
Show Events
}
