% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-spray_houses.R
\name{SprayHouses.none}
\alias{SprayHouses.none}
\title{No House Spraying}
\usage{
\method{SprayHouses}{none}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
The null model for house spraying
}
\keyword{internal}
