% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{Sugar}
\alias{Sugar}
\title{Sugar}
\usage{
Sugar(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{dependent variables vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Junction for sugar availability affecting adult mosquito
survival. Dynamic implementations are configured via \link{setup_sugar_object}
and activated by \link{dynamic_resources}.
}
\keyword{internal}
