% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_dynamics.R
\name{SugarDynamics}
\alias{SugarDynamics}
\title{Set the values of exogenous variables describing sugar}
\usage{
SugarDynamics(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a [list]}
}
\value{
[list]
}
\description{
This method dispatches on the type of `pars$SUGAR`.
}
