% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{Temperature.none}
\alias{Temperature.none}
\title{Set no temperature}
\usage{
\method{Temperature}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for temperature
}
\keyword{internal}
