% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{Temperature}
\alias{Temperature}
\title{Set the temperature}
\usage{
Temperature(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
Set the value of exogenous variables related to
temperature
}
