% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traps.R
\name{Traps}
\alias{Traps}
\title{Traps}
\usage{
Traps(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Junction for mosquito traps affecting blood feeding
(\code{Btraps}) and oviposition (\code{Qtraps}). Dynamic implementations
are configured via \link{setup_traps_object} and activated by \link{dynamic_resources}.
}
\keyword{internal}
