% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{Travel.dynamic}
\alias{Travel.dynamic}
\title{Travel (dynamic)}
\usage{
\method{Travel}{dynamic}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Dynamically updates time at home and travel EIR.
}
\keyword{internal}
