% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{Travel}
\alias{Travel}
\title{Travel}
\usage{
Travel(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Dispatches on \code{class(xds_obj$XY_interface$travel_obj)}
to update time at home and travel EIR.
}
\keyword{internal}
