% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{Travel.setup}
\alias{Travel.setup}
\title{Travel (setup)}
\usage{
\method{Travel}{setup}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Resets the \code{travel_obj} class to \code{static}, triggers
a setup update to the blood feeding model, and then calls
\link{travel_dynamics} once.
}
\keyword{internal}
