% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{Travel.static}
\alias{Travel.static}
\title{Travel (static)}
\usage{
\method{Travel}{static}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
For a static travel model, the function
does not update anything.
}
\keyword{internal}
