% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{travel_dynamics}
\alias{travel_dynamics}
\title{Travel Dynamics}
\usage{
travel_dynamics(t, y, xds_obj)
}
\arguments{
\item{t}{current time}

\item{y}{the state variable vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Computes and updates \code{time_at_home} and \code{travel_EIR}
on the \code{XY_interface} and \code{terms}, respectively.
}
\keyword{internal}
