% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-use.R
\name{Use_Bed_Net.none}
\alias{Use_Bed_Net.none}
\title{Null Bed Net Use Model}
\usage{
\method{Use_Bed_Net}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the \strong{\code{xds}} model object
}
\description{
Do nothing
}
\keyword{internal}
