% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-use.R
\name{Use_Bed_Net}
\alias{Use_Bed_Net}
\title{Use Bed Nets}
\usage{
Use_Bed_Net(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the \strong{\code{xds}} model object
}
\description{
Set the value of variables related to
bed net use
}
\keyword{internal}
