% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{Visiting.basic}
\alias{Visiting.basic}
\title{Visiting, the basic model}
\usage{
\method{Visiting}{basic}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a [list]}
}
\value{
a [list]
}
\description{
Implements [Visiting] for the basic model for Visitors
}
