% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{Visiting}
\alias{Visiting}
\title{Visiting}
\usage{
Visiting(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a [list]}
}
\value{
a [list]
}
\description{
This method dispatches on the type of `pars$VISITORS`.
}
