% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{visitor_dynamics}
\alias{visitor_dynamics}
\title{Visitor Dynamics}
\usage{
visitor_dynamics(t, y, xds_obj)
}
\arguments{
\item{t}{current time}

\item{y}{the state variable vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Computes and updates \code{visitors} (availability) and
\code{vis_kappa} (infectiousness) on the \code{XY_interface}.
}
\keyword{internal}
