% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{Visitors.dynamic}
\alias{Visitors.dynamic}
\title{Visitors (dynamic)}
\usage{
\method{Visitors}{dynamic}(t, y, xds_obj)
}
\arguments{
\item{t}{current time}

\item{y}{the state variable vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Dynamically updates visitor availability and
visitor infectiousness.
}
\keyword{internal}
