% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{Visitors}
\alias{Visitors}
\title{Visitors}
\usage{
Visitors(t, y, xds_obj)
}
\arguments{
\item{t}{current time}

\item{y}{the state variable vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Dispatches on \code{class(xds_obj$XY_interface$visitor_obj)}
to update visitor availability and visitor infectiousness.
}
\keyword{internal}
