% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{Visitors.setup}
\alias{Visitors.setup}
\title{Visitors (setup)}
\usage{
\method{Visitors}{setup}(t, y, xds_obj)
}
\arguments{
\item{t}{current time}

\item{y}{the state variable vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Resets the \code{visitor_obj} class to \code{static} and
then calls \link{visitor_dynamics} once.
}
\keyword{internal}
