% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{Visitors.static}
\alias{Visitors.static}
\title{Visitors (static)}
\usage{
\method{Visitors}{static}(t, y, xds_obj)
}
\arguments{
\item{t}{current time}

\item{y}{the state variable vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
For a static visitors model, the function
does not update anything.
}
\keyword{internal}
