% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{Weather.dynamic}
\alias{Weather.dynamic}
\title{Methods for exogenous variables describing weather}
\usage{
\method{Weather}{dynamic}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a [list]}
}
\value{
[list]
}
\description{
Implements exogenous forcing by [Weather]
}
