% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{Weather.none}
\alias{Weather.none}
\title{Methods for exogenous variables describing weather}
\usage{
\method{Weather}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements a null weather model
}
\keyword{internal}
