% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{xds_info_trivial_forcing}
\alias{xds_info_trivial_forcing}
\title{Forcing with Trivial Modules}
\description{
Trivial modules were developed for each one
of the three dynamical components, making it
possible to develop studies of some focal process
with known inputs: a \emph{trace function} approach.

Three trivial modules and EIR forcing
construct trace functions as composed time series. The
value of a forced variable \eqn{x(t)} is computed
as a product of four configurable elements:
\itemize{
\item \eqn{\bar x}: a mean value
\item \eqn{S(t)}: a seasonal pattern
\item \eqn{T(t)}: a trend
\item \eqn{K(t)}: a shock
}

\deqn{x(t) = \bar x \times S(t) \times T(t) \times K(t)}

In the trivial modules, the functions are specified by
passing parameters generated for \link[ramp.func:make_function]{ramp.func::make_function}.
\itemize{
\item \code{season_par} is passed to \link[ramp.func:make_function]{ramp.func::make_function} to compile \eqn{S(t)} or \code{F_season}
\item \code{trend_par} is passed to \link[ramp.func:make_function]{ramp.func::make_function} to compile \eqn{T(t)} or \code{F_trend}
\item \code{shock_par} is passed to \link[ramp.func:make_function]{ramp.func::make_function} to compile \eqn{K(t)} or \code{F_shock}
}

A trace function library is found in the satellite package \code{ramp.func}
}
