% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_sugar.R
\name{AvailableSugar}
\alias{AvailableSugar}
\title{Update the availability of sugar}
\usage{
AvailableSugar(pars)
}
\arguments{
\item{pars}{an [list]}
}
\value{
a [list]
}
\description{
This method dispatches on the type of `pars$SUGAR`.
}
