% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_sugar.R
\name{AvailableSugar.static}
\alias{AvailableSugar.static}
\title{Compute total availability of sugar}
\usage{
\method{AvailableSugar}{static}(pars)
}
\arguments{
\item{pars}{a [list]}
}
\value{
a [numeric] vector of length `nPatches`
}
\description{
Computes the availability of sugar for the static model (do nothing)
}
