% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation.R
\name{dynamic_importation}
\alias{dynamic_importation}
\title{Activate Dynamic Importation}
\usage{
dynamic_importation(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Activate dynamic importation by upgrading the
\code{importation_obj} to class \code{dynamic} and setting up the
travel and visitor sub-objects.
}
\keyword{internal}
