% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcing.R
\name{Forcing.dynamic}
\alias{Forcing.dynamic}
\title{Set the values of exogenous variables}
\usage{
Forcing.dynamic(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
With dynamic forcing, exogenous variables
can be set in one of four function calls:
- Weather
- Hydrology
- Shock
- Development
}
\seealso{
[dynamic_forcing]
}
