% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcing.R
\name{Forcing.setup}
\alias{Forcing.setup}
\title{The \code{setup} case for exogenous forcing}
\usage{
\method{Forcing}{setup}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Call all the functions to set the
values of exogenous variables and then revert
the \code{none} case
}
\keyword{internal}
