% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcing.R
\name{Forcing.setup}
\alias{Forcing.setup}
\title{The `setup` case for exogenous forcing}
\usage{
Forcing.setup(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
Call all the functions to set the
values of exogenous variables and then revert
the `none` case
}
