% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{Humidity.func}
\alias{Humidity.func}
\title{Set no humidity}
\usage{
\method{Humidity}{func}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for humidity
}
\keyword{internal}
