% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{Humidity}
\alias{Humidity}
\title{Set the humidity}
\usage{
Humidity(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
Set the value of exogenous variables related to
humidity
}
