% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{Hydrology.func}
\alias{Hydrology.func}
\title{Set no hydrology}
\usage{
\method{Hydrology}{func}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
[list]
}
\description{
The null model for hydrology
}
