% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-func.R
\name{make_bednet_coverage_function}
\alias{make_bednet_coverage_function}
\title{Set up dynamic forcing}
\usage{
make_bednet_coverage_function(
  options = list(),
  mean = 1,
  F_season = F_one,
  season_par = list(),
  F_trend = F_one,
  trend_par = list()
)
}
\arguments{
\item{options}{a list of options to override defaults}

\item{mean}{the mean bednet_coverage}

\item{F_season}{the seasonal signal in bednet coverage}

\item{season_par}{parameters to configure F_season}

\item{F_trend}{a temporal trend in bednet coverage}

\item{trend_par}{parameters to configure F_trend}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
\keyword{internal}
