% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_blood_hosts.R
\name{OtherBloodHosts}
\alias{OtherBloodHosts}
\title{Set the values of exogenous variables describing other blood hosts}
\usage{
OtherBloodHosts(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a [list]}
}
\value{
[list]
}
\description{
This method dispatches on the type of `pars$OTHER_BLOOD`.
}
