% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovitrap.R
\name{OviTraps.none}
\alias{OviTraps.none}
\title{Methods for oviposition traps}
\usage{
\method{OviTraps}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{OviTraps} for the none model (do nothing)
}
\keyword{internal}
