% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovitrap.R
\name{OviTraps}
\alias{OviTraps}
\title{Methods for oviposition traps}
\usage{
OviTraps(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{xds_obj$ovitraps}.
}
\keyword{internal}
