% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovitrap.R
\name{OviTraps}
\alias{OviTraps}
\title{Methods for oviposition traps}
\usage{
OviTraps(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a [list]}
}
\value{
[list]
}
\description{
This method dispatches on the type of `pars$ovitraps`.
}
