% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{Rainfall.none}
\alias{Rainfall.none}
\title{Set no rainfall}
\usage{
\method{Rainfall}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
[list]
}
\description{
The null model for rainfall
}
