% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{Rainfall}
\alias{Rainfall}
\title{Set the rainfall}
\usage{
Rainfall(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
rainfall
}
