% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{Resources.setup}
\alias{Resources.setup}
\title{Modify resources and resource availability}
\usage{
Resources.setup(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements [Resources] for the static model of resources
}
