% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{setup_development.func}
\alias{setup_development.func}
\title{Set up dynamic forcing}
\usage{
\method{setup_development}{func}(name, pars, Topts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an **`xds`** object}

\item{Topts}{a list of options to override defaults}
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
