% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{setup_development_func}
\alias{setup_development_func}
\title{Set up dynamic forcing}
\usage{
setup_development_func(
  pars,
  Topts = list(),
  mean = 30,
  F_season = F_flat,
  F_trend = F_flat
)
}
\arguments{
\item{pars}{an **`xds`** object}

\item{Topts}{a list of options to override defaults}

\item{mean}{the mean water level}

\item{F_season}{the seasonal signal in development}

\item{F_trend}{a temporal trend in development}
}
\value{
an **`xds`** object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
