% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{setup_humidity_func}
\alias{setup_humidity_func}
\title{Set up dynamic forcing}
\usage{
setup_humidity_func(
  pars,
  Topts = list(),
  mean = 80,
  F_season = F_flat,
  F_trend = F_flat
)
}
\arguments{
\item{pars}{an **`xds`** object}

\item{Topts}{a list of options to override defaults}

\item{mean}{the mean humidity}

\item{F_season}{the seasonal signal in humidity}

\item{F_trend}{a temporal trend in humidity}
}
\value{
an **`xds`** object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
