% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{setup_hydrology_func}
\alias{setup_hydrology_func}
\title{Set up dynamic forcing}
\usage{
setup_hydrology_func(
  pars,
  setup_no_forcing = list(),
  mean = 30,
  F_season = F_flat,
  F_trend = F_flat
)
}
\arguments{
\item{pars}{an **`xds`** object}

\item{setup_no_forcing}{a list of options to override defaults}

\item{mean}{the mean water level}

\item{F_season}{the seasonal signal in hydrology}

\item{F_trend}{a temporal trend in hydrology}
}
\value{
an **`xds`** object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
