% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{setup_no_development}
\alias{setup_no_development}
\title{Set up "no development"}
\usage{
setup_no_development(pars)
}
\arguments{
\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
Set up "no development"
}
