% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{setup_no_hydrology}
\alias{setup_no_hydrology}
\title{Set up "no hydrology"}
\usage{
setup_no_hydrology(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up "no hydrology"
}
\keyword{internal}
