% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovitrap.R
\name{setup_no_ovitraps}
\alias{setup_no_ovitraps}
\title{Set up the none model for oviposition traps (do nothing)}
\usage{
setup_no_ovitraps(xds_obj)
}
\arguments{
\item{xds_obj}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the none model for oviposition traps (do nothing)
}
