% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{setup_no_rainfall}
\alias{setup_no_rainfall}
\title{Set up "no rainfall"}
\usage{
setup_no_rainfall(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up "no rainfall"
}
