% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{setup_no_rainfall}
\alias{setup_no_rainfall}
\title{Set up "no rainfall"}
\usage{
setup_no_rainfall(pars)
}
\arguments{
\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
Set up "no rainfall"
}
