% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{setup_rainfall}
\alias{setup_rainfall}
\title{Set up dynamic forcing}
\usage{
setup_rainfall(Tname, pars, Topts = list())
}
\arguments{
\item{Tname}{the name of a model to set up}

\item{pars}{an **`xds`** object}

\item{Topts}{a list of options to override defaults}
}
\value{
an **`xds`** object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
