% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_season.R
\name{setup_season.none}
\alias{setup_season.none}
\title{Change season parameters}
\usage{
\method{setup_season}{none}(X, xds_obj, ix = 1)
}
\arguments{
\item{X}{a seasonality function object}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{ix}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Setup season for the NULL case
}
\keyword{internal}
