% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shock.R
\name{setup_shock.func}
\alias{setup_shock.func}
\title{Set up dynamic forcing}
\usage{
\method{setup_shock}{func}(name = "func", xds_obj, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{xds_obj}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
\keyword{internal}
