% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_dynamics.R
\name{setup_sugar_static}
\alias{setup_sugar_static}
\title{Make parameters for the static model for sugar (do nothing)}
\usage{
setup_sugar_static(xds_obj, Sugar = 0)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{Sugar}{describes sugar availability}
}
\value{
an \strong{\code{xds}} object
}
\description{
Make parameters for the static model for sugar (do nothing)
}
