% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{setup_visitors_basic}
\alias{setup_visitors_basic}
\title{Make parameters and functions for the basic model for visitors}
\usage{
setup_visitors_basic(
  pars,
  IMopts,
  x_scale = 0,
  xt = NULL,
  V_scale = 0,
  Vt = NULL
)
}
\arguments{
\item{pars}{a [list]}

\item{IMopts}{a [list]}

\item{x_scale}{a non-negative numeric value to set the mean for x_visitors}

\item{xt}{a function to change the pattern for x_visitors over time}

\item{V_scale}{a non-negative numeric value to set the mean availability of Visiting}

\item{Vt}{a function to set the temporal pattern for availability of Visiting}
}
\value{
[list]
}
\description{
Make parameters and functions for the basic model for visitors
}
