% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shock.R
\name{Shock.func}
\alias{Shock.func}
\title{Set no shock}
\usage{
\method{Shock}{func}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
[list]
}
\description{
The null model for shock
}
