% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shock.R
\name{Shock.none}
\alias{Shock.none}
\title{Set no shock}
\usage{
\method{Shock}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
The null model for shock
}
\keyword{internal}
