% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shock.R
\name{Shock}
\alias{Shock}
\title{System Shocks}
\usage{
Shock(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an **`xds`** object}
}
\value{
an **`xds`** object
}
\description{
Set the value of exogenous variables related to
systemic shocks
}
