% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{Weather.dynamic}
\alias{Weather.dynamic}
\title{Methods for exogenous variables describing weather}
\usage{
\method{Weather}{dynamic}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \link{list}}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements exogenous forcing by \link{Weather}
}
\keyword{internal}
