% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{Weather.none}
\alias{Weather.none}
\title{Methods for exogenous variables describing weather}
\usage{
\method{Weather}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a [list]}
}
\value{
[list]
}
\description{
Implements a null weather model
}
