% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{Weather}
\alias{Weather}
\title{Set the values of exogenous variables describing weather}
\usage{
Weather(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{xds_obj$weather}.
}
\keyword{internal}
